+++ CrownsGuard.Database/Ranked/RankedGame.cs
@@ -19,5 +19,5 @@ public class RankedGame
     public short Elo { get; set; }
     public bool IsHostStarting { get; set; }
 
-    public required byte[] Map { get; set; }
+    public required int[] Map { get; set; }
 }
